descartes_bi/requirements/development.txt
Werkzeug==0.9.3
django-extensions==1.1.1
